requirements.txt
asgiref==3.11.1
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
Django==6.0.2
gunicorn==25.1.0
packaging==26.0
pdfminer.six==20260107
pycparser==3.0
sqlparse==0.5.5
tzdata==2025.3
